// backend/src/types/systemSetting.types.ts
export interface SystemSettingInput {
  companyName: string;
  companyLogo?: string;
  favIcon?: string;
  phoneNumber: string;
  whtsappNumber: string;
  companyEmail: string;
  facebookUrl?: string;
  instagramUrl?: string;
  twitterUrl?: string;
  linkedinUrl?: string;
  websiteUrl?: string;
}

export interface SystemSetting {
  userId: string;
  companyName: string;
  companyLogo: string;
  favIcon: string;
  phoneNumber: string;
  whtsappNumber: string;
  companyEmail: string;
  facebookUrl: string;
  instagramUrl: string;
  twitterUrl: string;
  linkedinUrl: string;
  websiteUrl: string;
}